[fidlc] Informative failure on enum value generation
Currently, we do not validate the various literals used (be it for
constants, defaults, or enum members), opening the door to late stage
failures during code generation for things the compiler should catch
much earlier. While this doesn't fix the root cause of the issue (other
work is occuring to fix that), this at least provides a more informative
failure message.

FIDL-371 #progress

Test: crashes on a more informative way on the fidl file "library test; enum E { V = -1; };"
Change-Id: I997496aced3f541ada5728722abeb515e3dda801