[debugger] Changes to implicit component job.
Fixes a crash calling a null callback in the error case. The success
case null-checked the callback before issuing it but none of the error
cases did. I changed the callback to be required (removing the existing
null checks) and pass empty callbacks when they're not needed. This
seems less error-prone than remembering to null check before issuing any
callback.

Mark jobs disconnected when a connection is lost. This was the cause of
the error I was getting that caused the null callback described above:
when the connection is disconnected the job wasn't getting marked as
detached so when a reconnection happened, the job reported it couldn't
attach again.

Move implicit "component job" creation to the System object. It was in
the command-line frontend, and I think this feature is more of a core
debugger feature than one of the specific interface type.

Allows there to be no job contexts. Previously the ConsoleContext would
assert that there was always an active job. We will want to eventually
support deleting these so it's easy to get to no contexts.

Changes how the implicit job is reconnected. Previuosly the console
frontend would attach the current job context to the component job. But
the user could have switched to a different job context and the settings
of the current one might not apply. Now the implicit component job is
tracked and this job is automatically reconnected, retaining the
settings from a previous connection if there is any.

TEST=manual

Change-Id: Ia58dfcc25878a957cd8cd3d63f5e588262c485d3