[pkgfs] cleanup race conditions coming from GC
Test suites exist using the deprecated `pm install` flow that are performing
`pm install` during the GC that occurs 15s after pkgfs start. The GC
implementation introduced new "installing" metadata in the index that was
tracked based on blob fulfillment, rather than package activation. As such,
packages were only ever in the "installing" state from the GC perspective for
a short time - the time between the start and middle of "importPackage()".
All of the new tracking state is once again removed, as there was sufficient
state for all computation extant before those additions, and the extra
portions remain unused.

Additionally the GC run that occurs 15s after pkgs startup is removed, as
features enabling manual collection were written since the implementation. In
order to avoid challenges with OTA, the OTA process will trigger a GC before
starting to download packages.

The installation process for packages has the registration of the package
meta.far blob moved earlier in the process in order to eradicate the extant
race between the start of a package installation and when the packages set of
blob needs are identified.

Several state related races are removed or adjusted to eventual consistency
models, particularly during the installation process where errors were not
always propagated to callers.

Test: CQ (pmd_pkgfs_test)
Test: manual: pm install
Test: manual: unlink /pkgfs/garbage
Bug: PKG-364 #done
Change-Id: I3ee18cca02795f893ca06c8c16a3d4bf20cafe70