[zx][topology] Add core system-topology logic.
This defines both the "flat" form of the topology that we expect to be
passed from the bootloader or generated in early boot. It also defines
the "heap" form of the topology that will be maintained at runtime and
referenced by all interested parties.

This currently is static after early boot and as such doesn't use locks,
if we move to a world where we want to update this later in execution
(when we are multithreaded) we will need to add locks.

Issue: ZX-3068

Test: Added hosttest unit tests.
Change-Id: Ie2490a1d8ffa5b4fffa5531b0ee71118e0075279