[callback] Make ScopedTaskRunner thread-safe
This augments ScopedTaskRunner to allow for thread-safe implementations, and
adds a default implementation that is unsynchronized but allows callers to
use it in a thread-safe way.

A ScopedTaskRunner based on fxl::WeakPtr is not thread-safe because
fxl::WeakPtrFactory does flag twiddling and does not allow for a shut-down
state, which can lead to race conditions when tasks are posted from a thread
other than the dispatch thread.

Also adds more unit tests.

Depends-On: https://fuchsia-review.googlesource.com/c/peridot/+/229101
Test: fx shell runtests -t garnet_lib_unittests
Change-Id: Ie9f1b2ee7b77cdadb0f94f5200311b8bd527ad07