[zircon][usb-request]Add node_offset field to usb_request_pool
usb_request_pool is a pool of usb_requests that a driver keeps track
of, to avoid reallocation of requests. With the recent changes to usb stack,
each driver in the usb stack has its own list_node_t* so as to avoid stamping
on other driver's lists. The library is unaware of the private portions of the
usb_request_t where the node pointer is stored for each driver. To queue
the usb_requests to the pool, this changeset informs the
usb-request library, the offset of list_node_t* from the
usb_request_t.

Test: vim2: fx serve. lsusb, iochk on mass storage device.
Astro: lsusb.
Pixelbook: fx serve. iochk

Change-Id: I78ea612472742ad93b6e5c61ffe09da7819276fd