[syscalls] Add rx_buf_available to zx_info_socket
We currently report the total amount of data that clients can ever read
out of the socket, but we need to know the amount of data that can be
read out of the socket in one zx_socket_read call in order to replace
the NULL/0 behavior from zx_socket_read.

This CL adds rx_buf_available to zx_info_socket to contain that
information.

Test: socket_buffer_test
Change-Id: Icb8f2bc314eed5d4afce6190fd11cc08ef6b9d00