[process builder] use the resolver.
Adds support to the ProcessBuilder for the Resolver service. This
checks for the #!resolve prefix in an executable and queries the
resolver for the proper destination and loader.

This logic is adapted from the FDIO spawn library.

Remove the CloadLdsvc function from the ProcessBuilder. When one loads
via LoadPath, one will almost always want the loader service that was
resolved. When loading a specific VMO, I think it will be common to
specify an override (there are currently no users of this function so
it's hard to make good decisions about what its callers want).

If we need the ability to override the loader service for
ProcessBuilder, we can add that back later.

TEST=manual

Change-Id: Ib3d8bf0a5e2d5051039e70ed7e916660f44d3ca7