[mediaplayer] separate command queue from cq tests
This CL factors out the command queue previously added to the cq test
code so it can be used for manual tests as well. The command queue is
useful for constructing automated repros, and we need to be able to
run those repros in a real environment, not just the hermetic
environment.

TEST: test change, CQ still passes
Change-Id: I3d8d4bb98bb326434d6c3e7be76b87095a7f188e