[rights] Add MANAGE_PROCESS to DEFAULT_JOB, MANAGE_THREAD to DEFAULT_PROCESS
Add PROCESS management to jobs by default, add THREAD management to
processes by default. The logic here is that job handles need to be able
to create processes inside themselves, and that process need to be able
to create threads inside themselves.

Additionally, accept MANAGE_PROCESS on a job as sufficient to
process_create() with a fallback to WRITE for all the exisiting callers
that are calling with WRITE.

ZX-2967 #comment [rights] Add MANAGE_PROCESS to DEFAULT_JOB, MANAGE_THREAD to DEFAULT_PROCESS

Test: CQ including garnet and topaz
Change-Id: Icf767a369299596c2d5974225fb9e31fa59a923b