[mediaplayer] allow flushed->paused transition in all cases
The implementation of the Pause method did nothing when 'can pause'
was false. When 'can pause' is false, the player cannot be paused
once it's playing. The Pause method is often invoked immediately
after opening new content in order to queue up the content for
playback. This should be permitted regardless of 'can pause'. Also,
because the Pause method implementation is pipelined, Pause is
frequently called before the player knows that 'can pause' should
be true.

TEST: tested manually using mediaplayer_test_util
Change-Id: I73082a295d897a757eab29a46493d99bcfa8bb83