[mdns] fix name compression exploit
Certain input datagrams can crash mdns, because it allows loops to
occur in name compression chains. This CL addresses that by ensuring
that names never exceed their overall maximum length of 255 bytes.

The fuzzer that uncovered this bug also revealed that the parser
wasn't checking properly for the two bits that need to be set in
a name label length to indicate a compression offset. This has
been fixed as well.

TEST: manually tested using avahi

NET-1962 #done

Change-Id: I6f8d0150a912b1a8c1e903397cd7c80915515ce7