[netstack] Register waiter on TCP before writing
If the write happens before the waiter is registered, the inbound
packet might be lost.  This causes a deadlock in the test.  The solution is to register before writing.

TESTING:
This command no longer hangs the shell:
`total=0; while [ $total -lt 100 ]; do ./netstack_link_bridge_test;  total=$((total+1)); done`

NET-1957

Change-Id: I1b0d5bceda6e34cf4bb6106fa431a04cbadaf528