[mediaplayer] use command queue in mediaplayer_test_util
This CL integrates the command queue, which was previously factored
out of the CQ test, into mediaplayer_test_util. mediaplayer_test_util
is used for manual testing. The existing features of the utility have
been reimplemented using the queue. In addition, a new --experiment
option has been added for developers to drop in ad-hoc experiments.
Of course, new features are bug repros can always be added as needed.

TEST: manually tested (this is a manual test app)
Change-Id: I399f4c82c4216d0fe1540eb50fa6efd6af5f1cbf