[tracing] Introduce duration complete events.
This CL introduces complete events which is an event that contains
both its start and end time. This is meant to be used for events
that have identical start and end information and they only differ
in the timestamps.
This should save upwards of 45% on trace objects that uses.
It would also save similar % in the output format.
(Which is implemented in a followup CL)

This CL does not add any real usage (beside the testing) of the new
Event Type to avoid breaking tracing (since the exporter is not yet
implemented).

A follow-up CL in zircon will introduce the usage.

TEST: added tests, running on CQ + manual testing with the rest of the
ongoing CLs.

BUG: PT-72

Change-Id: I3bfd023c09022e0cb6e15f6bb090f7a612e0171a