[usb-test-fwloader] FX3 initial setup part 1.
The default FX3 bootloader has a bug, so we want to load a second
stage bootloader to the I2C EEPROM when setting up a new test device.

The first step is to load the FX3 Flash Programmer image to RAM,
which provides a vendor command to write to the EEPROM.

ZX-3045 #comment

TEST= copy flash programmer image
usb-test-fwloader -e -i firmware.img -p flash.img
lsusb and confirm PID 0x4720 shows up.

Change-Id: Ide2f3e1a1c00411e237840ad9219a694a740b2ac