[tftp] track paving state through errors
Always wait for the paver process to exit, regardless of state - executing
the paver concurrently has undefined behavior.

If the paver has received all data from the file transfer, capture the return
code, and on non-zero, propagate that code back to bootserver.

The layout of this patch is not ideal - it would be preferrable to track the
state in the file_info field, however, the asynchrony and protocol result in
a requirement to check for the error inside of "future" nodes in the state
machine (i.e. on the next file, rather than the file that fails) under many
conditions. The exit code must be checked at the first opportune moment,
propagated and reset at that point.

Test: manual: created a new fvm.sparse.blk containing only the first 50 blocks of a valid one. Old code continued after error, new code reports error to bootserver.
Bug: ZX-2996 #comment zedboot reports paver errors to bootserver
Change-Id: I9e191a5dbe0f744e7e86d00976330bfe6fe5d978