[arm][gic] Change EOImode to separate deactivation
Set the EOImode so that EOIR only drops the priority, and DIR is
required to deactivate the interrupt. This allows the hypervisor to
correctly take control of physical interrupts, without affecting
operation of the host.

In addition, this updates the boot-shim for mt8167s to enable the use of
GICC_DIR, which is located 64k from the GICC base. The first page of
GICC is mapped repeatedly into the the first 64k.

MAC-180

Test: Test: Ran on a VIM2 (for GICv2) and QEMU (for GICv3).
Change-Id: I7fd2baaa2d230f6652ced32a02807f0d8fad2f79