[block] Re-write block FIFO thread as ServerManager
- Drastically simplify lifetime management of the FIFO
server, removing a "lock, completion signal, threadcount,
and boolean" with a ServerManager class that provides
a simple API.
- ServerManager utilizes a thrd_t and an atomic to identify
if a FIFO server is running, and if so, in what state. The
most notable improvement comes from ServerManager's ownership
of the background thread, rather than detaching it after
launching.
- Isolating this logic into a "ServerManager" class will
make this API trivial to unit-test, as soon as the "blockserver_*"
APIs are abstracted.

Test: /boot/test/sys/ramdisk-test

Change-Id: I07c951db31c663b0aab68da2794b77408723ef07