[debugger] Start to separate out variable finding.
Previously variables were found at the same time their value was
computed. This ends up adding complexity and making some code paths
asynchronous when they could be synchronous.

The plan is that there will be a two-phase variable lookup. In the first
phase, the variable is found, and in the second, the value of that
variable is computed. This will also allow future flexibility in terms
of autocomplete and in looking up symbol information for different types
of variables.

This patch starts this process. It separates out the code to resolve
members of |this| and local variables, and the SymbolEvalContext
converts these two values separately.

In subsequent passes this logic will be further split and the
SymbolEvalContext will be relatively small. Currently, there should be
no behavior change.

TEST=unit

Change-Id: I8ebb45527cda5d6680bad4c33751def41bb2f744