[wav_recorder] Move out of examples
wav_recorder has grown to be a tool rather than a simple example
of how to capture a wav file.  Move it from examples/media to
bin/media and make it part of the standard bits included from
garnet.

In addition, un-componentize it as we depend on being able to
grab standard input, which is currently unavailable for interactive
shell processes that are componentized.

TEST: build+pave, verify wav_recorder exists and works.
Change-Id: Ib38592809f3b242e07b241b955db103ea0111fc3