[bt][host][gatt] Fix stack to DDK error conversion
The "protocol_error()" field of a core stack Status object has an
undefined value if it doesn't represent a protocol error. The GATT
service bus incorrectly interpreted this which caused intermittent
failures in GATT interactions. In addition, the generic "status"
field of an error was always getting set to ZX_OK, ignoring internal
errors that can occur in the absence of a protocol error.

The bus now correctly converts both internal and protocol error fields
of a btlib::att::Status when creating a DDK bt_gatt_err_t.

Bug: BT-645
Test: 1. Pair a BLE keyboard that has a report map characteristic.
      2. Disconnect and auto-reconnect the keyboard.
      3. The bus driver should not fail to read the report map
         characteristic. The keyboard should be usable.

Change-Id: If2519da4cb6d033cb932955f39c5d9ba1237fab2