[netstack] Manually serialize/deserialize MacAddr
This change implements the serde::Serialize and serde::Deserialize
traits for MacAddress manually so that the struct is serialized as a
string. By doing so, the MacAddress can now be used as the key of a JSON
object.

NET-1302 #progress

TESTED=fx run-test fidl_zircon_ethernet_ext_lib_test_rustc

Change-Id: Ib0181286c9a21163c9ca8b07e4f6324d14ef843c