[tracing] Stop splitting bounded duration events into start and end events
This is step 3, where is switch the calls that already uses bounded
trace event to use the duration complete events.

This will transform the kernel syscalls and the vcpu events to
the new duration events, almost halving the size needed for those
in the trace buffer and in the output file.

The next step is to make more calls use the bounded event records
mainly the TRACE_DURATION macro.

BUG: PT-72

TEST: manually verifying that those events appear correctly in traces +
CQ

Change-Id: I6f8a71d281845b431077b914541565cd32065e0f