[guest] Fix device initialisation
The call to Start should be synchronous. Even though we currently use a
SyncPtr, we need to return a result to generate synchronous code. The
current generated code calls SynchronousProxy::Send, which does not
block. After this change we will generate code that uses Call, which
does block.

TEST=Ran guest_integration_tests in a loop for a while. And
device_tests.

Change-Id: I9a596dde53e3cf6c9780f299d767290daed52026