[upci] Implement the upstream concept in userspace
In kernel PCI, an upstream node was an interface
for bridges and roots to allow for downstream allocation,
and hooking scanned devices into the topology. In the user
PCI world bus scanning and ownership will be handled by
the top level Bus class, but we still need to have a hierarchy
of allocators for device space downsteam. Unlike kernel PCI,
this version of upstream implements two kinds of allocators:

1. Root allocators which allocate address space over the Pciroot
   protocol. This type is used by Root nodes.
2. Region allocators which allocate address space from their own
   region allocators which were allocated from the node upstream
   of them.

Some method signatures have been included but are not implemented.
These may go away in the future as the Bus driver interface is
finalized. The bulk of the UpstreamNode interface related to downstream
devices will be added back to the class as pci::Device objects are
ported over to userspace to avoid CLs spanning thousands of lines.

ZX-3146

Test: Compiled and ran upci, no regressions detected.

Change-Id: I95dc56cab4ef6a1672c5eee7424a1e1995ba5f36