[bt][gatt] Server support for service-level security
The server data bearer now generates the "Insufficient Authentication"
in response to read and write request types based on the security properties
of the link and the requirements of the attribute in request.

Enforcement is currently done by gatt::Server in several request
handlers. We should aim to reduce the points of failure by moving the
enforcement logic to centralized location in att::Attribute or att::Database.

BT-157 #done
Test: bt-host-unittest --gtest_filter=GATT_ServerTest_Security.*

Change-Id: I3dfe399d5ce5a01cf4040283239024b03b51a771