[fdio] do not accept sockets with a full fd table
If the fd table is full, pulling sockets from the control channel results in
consuming the incoming connection, but being unable to actually service it -
we then immediately close (which then ends up in linger). Other processes
would then not be able to pull these sockets (in a load balance scenario),
and netstack also potentially loses the ability to apply policy from having a
full listen queue. Instead, try to allocate an fd up front, and only if found
accept the socket from the control channel. This is more conformant with
common accept behaviors.

Test: install a package with 5500 blobs with a package server routed via a remote port forward over ssh. Observe successful install rather than failure.
Bug: PKG-371 #comment fdio: do not accept sockets with a full fd table
Change-Id: Idb5513ada590b0d8da2f5cd68b4d734765e438d4