[resolve] Trampolines end in newline.
The process launcher expects #!resolve VMOs to end in a newline.

Previously, this happened to work because the VMO was padded with \0's,
which meant that the resource name passed to the Resolve call was
interpreted as the right string (albeit with the wrong length).

However, for other languages that do not use C-strings (such as Rust),
this causes problems, particularly when the string is checked when
passed through a C ABI boundary.

TESTED=logged that the resource name had the right length

Change-Id: I6ed94c24bf446958d0546c1d96cb5f7a06aa93fb