[bt][test] Re-enable host_lifecycle and host_integration tests
The Bluetooth host_lifecycle and host_integration tests were disabled as
they were flaking due to an occasional HCI timeout. The timeout is being
investigated further as BT-604. We will also investigate a better way of
recovering from such rare timeouts as BT-603. In the meantime, we have
increased the timeout and this should allow us to re-enable the tests
without them flaking on CQ.

US-557

Test: The two sets of tests re-enabled by this commit
Change-Id: I4ac131966a2648dc28e446abe3609bfd3b87d9d5