[elflib] Add GetMappedMemory to MemoryAccessor Interface
When accessing ELF data in a running process we may need the mapped
address, not the offset, to get the data we're looking for. This allows
the MemoryAccessor interface to service such requests correctly.

While we're in here, clean up some of the extraneous helper functions we
have lying around.

Test: Modified unit tests pass.
Change-Id: I3d2ed9d08fae6765e081744d3f9e5e3ccdcda430