[mediplayer] provide sample format to SBC decoder
This CL supplies a planar sample format to the SBC decoder, which the
decoder will copy to its output format. Previously, sample format
AV_SAMPLE_FMT_NONE was sent, causing the buildout of the graph to
fail.

TEST: Tested with modified CQ test. CQ test will be added.

Note on testing: The failure mode associated with the issue fixed
here is problematic. When the failure mode is fixed (i.e. the
client is told the format won't work), a CQ test for this issue
will be added.

Change-Id: I08546919eab02dbb32aa6c0d7d1bb2e671af483f