[inspect] Implement inspect VMO reader.
The new Reader takes a VMO Snapshot and parses it, returning the root of
a tree of Objects in the VMO.

This reader exists in Garnet, since it uses C++14 types and the standard
library.

This will eventually be merged with the general reader for the facade.

This change additionally removes path storage from the reader interface
in favor of allowing clients to keep their own state.

TEST=runtests
CF-216: #done

Change-Id: I288164e68f70e4053d8145493a0d31fbbdb94924