[debugger] Make the LLVM line table mockable.
Adds a line table interface in between our implementation and LLVM. This
will allow mocking of the line table in future tests.

This does not currentl change any behavior or add any tests. The line
table lookup logic will be separated out to its own file in a subsequent
pass and tests will be added for that file using the mock line table.

TEST=this is test support

Change-Id: I9f7ceec23c5d8ab69aa373d93852655d247d66cf