[rust-netstack] 1-index DeviceIds
The fuchsia.net.stack API does not allow zero as a valid device ID. In
order to prevent zero being used, the change defaults the next available
ID in the DeviceLayerState to 1. If we want a stronger guarantee, we
could also change the ID allocation function to not allow allocating
zero as an ID. However, for now, I am preferring the flexibility of
using default.

Test: None
Change-Id: Idfcb7729eb34360d2d31c7d75af6ca40c6aab38e