[rust-netstack] Refactor EventLoop to use mpsc queue
This refactors the EventLoop to use a single mpsc queue for all events.
This way, the event loop is simply a loop that drains the queue,
matching on the type of event.

This allows us to get rid of the Arc<Mutex<Context>>, since now, if
something needs the outer EventLoop state from within the
EventLoopInner, instead of having to have ownership over the Context, it
can now have a cloned mpsc Sender, which allows it to send a message to
the main event loop when some action needs to be taken that requires use
of the Context.

Now, everything that needs to trigger an event is a future that has a
mpsc sender, allowing it to send messages to the main event loop when it
wants to trigger an event. Currently, the events are FIDL, ethernet
events, ethernet setup events, and timers firing.

By having a single queue for events, instead of a queue per event type,
we can guarantee FIFO ordering for messages, which makes the system much
easier to reason about.

This change also adds the MVP implementation of the
fuchsia.net.stack.Stack FIDL API required to test this change.

This change affects ping performance in two ways - it increases average
latency on Toulouse by 0.025ms, while dramatically reducing variance.
For a test of 100000 pings, min/avg/max/mdev is 0.171/0.225/9.571/0.044
ms before this patch, and 0.201/0.255/1.300/0.017 ms after this patch.
The performance difference in QEMU is similar, although latency and
variance are worse across the board.

Ping RTT histograms for Toulouse:
* before: https://drive.google.com/open?id=1-DAvDvZmjPhZpSaJRkgKOhtfKG8RUnv3
* after: https://drive.google.com/open?id=1gZOt_CUx-4fwvCnXmohhSqlVX31CVdZJ

Test: Manually tested. Steps:
Add recovery netstack to sysmgr, and disable go netstack
qemu$ run net-cli if add /dev/class/ethernet/000
qemu$ run net-cli if addr add 1 192.168.1.39 24
qemu$ run net-cli fwd add-device 1 192.168.1.0 24
host$ ping -I qemu 192.168.1.39 -c 10

Change-Id: I25bbcd3a3135a123d4c43dc43112a226f811ea7b