[bt][drivers] Update the usb request header length correctly
The data transfer size in a usb request is conveyed to the lower layes
via the header.length field. The atheros driver creates one usb request
of 4096 bytes and uses it to load the firmware. The last transaction transfers
less data than the allocated request size. In this case, currently the
req->header.length is not updated to the exact amount of data that is being
transferred and so the request does not get completed. The atheros driver infinitely
waits on the completion of the last request. The firmware load happens in the bind
thread and the bind thread is now blocked infinitely. I think this is the reason the
whole usb stack is blocked, because communication with devmgr is now blocked. Ideally
I/O intensive tasks like loading firmware should happen on a separate thread that is
spawned from the bind thread, but that is outside the scope of this
changeset and I will raise a separate bluetooth bug for that. What I am
still unclear is, why this popped up only recently. This should have always
been an issue.

ZX-3107 #Done
WLAN-98 #Done

Test: fx full-build. Without this change, the firmware load fails on AVA. lsusb
is stuck. fx shell is stuck. With this change, all the above work, Ofcourse with the
btatheros stack. I see logs: "btatheros: loaded successfully" and all the
usb devices coming up later.

Change-Id: Ie9607a21a9350b1a68e63eaac4d484c8e04bd83d