[modular tests] Use --session_shell in tests.
In a test, the presence of a session shell config file does
not override the --session_shell command line flag value.

This allows to run the integration tests even in builds from
product configurations with session shell files. In an
integration test, the session shell is never taken from a file,
and this works only because integration tests are usually run
in build configurations without a product.

TESTED=run_modular_tests.sh passed and does not time out.

Change-Id: I63a8e4a163f2ae29c0fd5aaa8e7d32b41282388b