[amber] Separate source loading from monitoring
Sources can either be enabled or disabled, but, in a few cases, it was
possible for different instances of the source to exist in both states
simultaneously, and it was possible for an enabled source to have more
than one instance of its background tasks running.

This change ensures that if a source is being removed or disabled, its
background tasks are properly cleaned up. This change also moves the
responsibility of spawning a sources background tasks from the source
itself to the daemon, allowing a source to be loaded without
automatically spawning its background tasks.

Test=various combinations of add/remove/enable/disable sources with
extra monitoring in place to ensure source watchers were not started
when they shouldn't be and that they were cleaned up when needed.

PKG-375 #done

Change-Id: I5926acafb90cd6955d1231f51eb945af84bbfec5