[fit][bridge] Fix a lock re-entrance issue in fit::bridge.
Under certain circumstances, dropping a bridge completer or consumer
could cause a result and/or suspended task to be dropped which could
then have side-effects that resulted in an attempt to re-acquire the
bridge's lock.

Guarded against this by ensuring that the result and suspended task
are only used or dropped from outside of the bridge's lock.

Added swap() functions to fit::result<> and fit::suspended_task to
help implement this fix.

Test: fit-test
Change-Id: I046c71b8f37eb3242a8bb2a5657986bab75a154f