[debugger] HW breakpoint test
NOTE: The testing for arm64 is disabled until the debug agent supports
setting those breakpoint in.

This CL extracts a lot of the common functionality used for the SW
breakpoint classes and separates them into their own objects so that
they can be more easily reused in different tests.

This will make testing of other constructs, such as watchpoints, much
easier to do.

This also adds a test to verify that there is a HW breakpoint
notification in place.

TEST=Unit

Change-Id: I5525d1a6ddb89de144683d65acf7a23cb36cfa24