[mediaplayer] don't crash on unknown stream type
This CL fixes the ConversionPipelineBuilder so it no longer crashes
when a stream type cannot be handled properly. This class was coded
to crash the process in order to aid debugging during early
development. Now we need to handle failure cases smoothly.

This CL also changes the way candidate stream types are rated,
acknowledging that some audio types cannot be supported at all. In
those cases, the candidate stream types are given a 0 rating.

TEST: tested with new CQ test to be submitted in another CL
Change-Id: Iae07c3cd1a5e2a79fe9d1a48faa450c14f0854f7