[mediaplayer] improve status reporting
This CL addresses status reporting issues that cause confusion for
the client. In one case, a status update was not provided when a
stream was connected, preventing the client from being informed
that the connection was established. In another case, connection
success was indicated prematurely and contradicted later, creating
false results for tests.

TEST: tested with new CQ tests to be submitted in another CL
Change-Id: I87235bdd0580adb677d5d291d5d9ada79984de1b