[debugger] Separate out some line searching.
Moves some logic for line searching and matching from the ModuleSymbols
implementation to a new file.

Adds a new function to the LineTable to lookup the subroutine for a line
table row. This allows the new helper functions to avoid using the
DWARFUnit completely.

The unused "unit" member of the LineMatch structure was removed.

There is no behavior change, code move only.

TEST=unit,manual

Change-Id: Ib8847de8fcd09c631e0bc3920ff7d898b57be48d