[scenic] Simplify Command Buffer Sequencer
Command buffer sequencer had two helper classes to attempt to limit access to various subparts of the interface. The former was not necessary (and it limits the ability for classes to use encapsulation instead of inheritance with AddListener) and the latter has been replaced by friending the appropriate classes (CommandBufferPool and the necessary unit tests).

In addition, this code fixes SCN-497, as DescriptorSetPool is now a ResourceRecycler, instead of simply a ResourceManager.

tested=scenic_unittests, default UX flow

Change-Id: Ic4542477f0424476fbca4272a11279276e8dedb5