[usb proto] Make complete_cb as a parameter to usb_request_queue
Already reviewed at
https://fuchsia-review.googlesource.com/c/zircon/+/232612
However, when including that change in integration review test with
garnet change, (patches.json), the dry run fails,
unable to apply the zircon change.Infra team is also unsure about the
reason. I synched to JIRI_HEAD and removed all the associated
dependencies, and generated this change with exact same source and
the dry run now succeeds.

With usb request having private regions for each layer in the stack,
the complete callback and their contexts should not be part of the public
part of the usb-request. This changeset modifies usb protocol so that
the complete_cb and its context are passed in as parameters.

Test: Test with garnet change: https://fuchsia-review.googlesource.com/c/garnet/+/232701
pixelbook: fx serve. fx shell. lsusb. iochk on ums device. audio -d 1 tone.
Astro: netboot. lsusb.

Change-Id: I691ca2ccd8eeca35711d501ac7b23be4d91161dd