[fdio] only accept sockets when there's a free fd
This change introduces a reservation mechanism throughout fdio that can be
used in the intermediate time between a callers request for a new fd, and
when that request is fulfilled by some remote. This reservation mechanism is
then used in the socket accept path to prevent client programs from pulling
sockets from the listen queue if they do not have an fd with which they can
service the socket.

A prior change was reverted in bbdb4a51d5849d91c73945994be75db397528dd5 which
instead attempted to simply hold the fdtab mutex for the duration. That
approach was flawed because it creates a deadlock in the (common) socket
scenario of:
- listen
- thread(connect)
- accept
If the accept call races connect.

This approach likely should also be used in open and connect among other
places to avoid creating new remote resources that can not be serviced.

Test: garnet package: netstack_tests
Bug: PKG-371 #comment accept does not pull sockets it can't service
Change-Id: Ic7745bca80c6cec3bad468bfc1d923a275e4159b