[tracing] Depend on libtrace when that is what's really meant
libtrace-provider has libtrace listed as a dependency when really it
does not have any such dependency. This lets trace clients use
libtrace-provider as a dependency when what they really mean is libtrace.

This errant dependency in trace-provider is being fixed, which
means we need to fix these clients.

PT-79 #comment patch

Tested: CQ
Change-Id: Ic0433516f8b7bf4f83951ab3dca60a00a0efa704