[usb protocol] Send complete callback and context as parameters
This is already reviewed at
https://fuchsia-review.googlesource.com/c/garnet/+/232320.
However, when including that change in integration review test with
zircon change, (patches.json), the dry run fails
failing to apply the zircon change.Infra team is also unsure about the
reason. I synched to JIRI_HEAD and removed all the associated
dependencies, and generated this change with exact same source and
the dry run now succeeds.

With usb request having private regions for each layer in the stack,
the complete callback and their contexts should not be part of the public
part of the usb-request. This changeset modifies usb protocol so that
the complete_cb and its context are passed in as parameters.

Test: with zircon changehttps://fuchsia-review.googlesource.com/c/zircon/+/232759:
pixelbook: fx serve. fx shell. lsusb. iochk on ums device. audio -d 1 tone.
Astro: netboot. lsusb.

Change-Id: Id85d382b25eb96927c29f08eeffb4f3a477b04b5