[benchmark/story] Add libtrace dependency
libtrace-provider has libtrace listed as a dependency when really it
does not have any such dependency. This lets trace clients use
libtrace-provider as a dependency when what they really mean is libtrace.

This errant dependency in trace-provider is being fixed, which
means we need to fix these clients.

PT-79 #comment patch

Tested: CQ
Change-Id: I58fde0d6a8814a32c93d9385d10157dd46fcd52a