[debugger] Improve line matching.
Previously we found the first match in a function before computing the
best match. This gets confused if the code is reordered in a function
where the best match is not the first time the line table crosses the
line in question (in crashpad_analyzer in the bug, the functions were
getting reordered).

Now the best match is found across all functions before the function
filtering is applied. Tests are added for this case.

DX-755
TEST=unit

Change-Id: I60c0b5b5ae4f868170cc0dd2398cac800509a01f