[sl4f] add wlan facade base
Adds the structure to support wlan in sl4f.  This does not add new wlan
tests, but instead creates the framework for wlan tests.  The first
test, scan and verify there are results, will be added in a follow-on
cl.

Used a corresponding CL in acts that triggers a scan test, to verify
plumbing for scan commands.  The test is not currently active.

Bug: WLAN-529
Test: act.py -c basic.config -tc WlanScanTest (and manually verified
logs that confirmed the scan request was received)

Change-Id: I8da5143460012fa52dac575815f177b593866451