[debugger] Improve line lookup
Previously the line lookup would look for transitions from a line before
the one being searched for to one greater-than-or-equal. This missed
some matches. If a line after the one being searched for was reordered
to right before the best match, the transition would be in the opposite
direction.

This patch uses a simpler algorithm. It just finds the smallest line >=
the one being searched for.

DX-755
TEST=unit,manual

Change-Id: Iaf851fc21f7edfa5d55a6e7b7363fe98ab7b5aff